Remove jquery
I'd added it when trying to show the dropzone div. Turns out the problem
was with turbolinks. So my new solution was to disable turbolinks on
the link to that page.

The other solution is to initialize the dropzone on the page switch
event rather than on the page load event.